+++ packages/server/src/index.ts
@@ -1,4 +1,5 @@
 export {Server} from './server'
 export * from './types'
 export * from './models'
+export * from './lockers'
 export * from './constants'
