+++ packages/server/src/models/Locker.ts
@@ -1,106 +1,24 @@
-import {ERRORS} from '../constants'
-
 export type RequestRelease = () => Promise<void> | void
 
-// The Locker interface defines methods for implementing a locking mechanism.
-// This is crucial for ensuring exclusive access to uploads and their metadata.
-// Following TUS recommendations, it's important to cancel locks from previous requests
-// to avoid holding locks for too long and to manage half-open TCP connections effectively.
-// The lock method includes a cancel callback to facilitate the cancellation of a request that previously acquired the lock.
+/**
+ * The Locker interface defines methods for implementing a locking mechanism.
+ * It is primarily used to ensure exclusive access to resources, such as uploads and their metadata.
+ *
+ * The interface adheres to TUS protocol recommendations, emphasizing the need to prevent prolonged lock retention.
+ * This approach helps manage resources efficiently and avoids issues with half-open TCP connections.
+ *
+ * Methods:
+ * - lock(id, cancelReq): Acquires a lock on a resource identified by 'id'. If the lock is already held by another request,
+ *   the 'cancelReq' callback is provided to signal the current lock holder to release the lock.
+ *   The 'cancelReq' callback should be invoked when there's an attempt by another request to acquire a previously locked resource.
+ *   This mechanism ensures that locks are held only as long as necessary and are promptly released for other requests.
+ *
+ * - unlock(id): Releases the lock held on the resource identified by 'id'. This should be called by the lock holder
+ *   after completing their operation or upon receiving a signal through the 'cancelReq' callback from a subsequent request
+ *   attempting to acquire the lock.
+ *
+ */
 export interface Locker {
   lock(id: string, cancelReq: RequestRelease): Promise<void>
   unlock(id: string): Promise<void>
-}
-
-export interface MemoryLockerOptions {
-  acquireLockTimeout: number
-}
-
-class Lock {
-  public requestRelease?: RequestRelease
-}
-
-// MemoryLocker is an implementation of the Locker interface, maintaining locks in memory.
-// It ensures exclusive access to upload resources by managing locks for each upload.
-// The lock() method ensures that any previously held lock is released before acquiring a new one.
-// Lock acquisition attempts will timeout based on the specified 'acquireLockTimeout' duration.
-export class MemoryLocker implements Locker {
-  private locks = new Map<string, Lock>()
-  protected timeout: number
-
-  constructor(options?: MemoryLockerOptions) {
-    this.timeout = options?.acquireLockTimeout ?? 1000 * 30
-  }
-
-  async lock(id: string, requestRelease: RequestRelease): Promise<void> {
-    const abortController = new AbortController()
-    const lock = await Promise.race([
-      this.waitTimeout(abortController.signal),
-      this.acquireLock(id, abortController.signal),
-    ])
-
-    abortController.abort()
-
-    if (!lock) {
-      throw ERRORS.ERR_LOCK_TIMEOUT
-    }
-    lock.requestRelease = requestRelease
-  }
-
-  protected async acquireLock(id: string, signal: AbortSignal): Promise<Lock | void> {
-    if (signal.aborted) {
-      return
-    }
-
-    const lock = this.locks.get(id)
-
-    if (!lock) {
-      const lock = new Lock()
-      this.locks.set(id, lock)
-      return lock
-    }
-
-    await lock.requestRelease?.()
-
-    const potentialNewLock = this.locks.get(id)
-    if (!potentialNewLock) {
-      const lock = new Lock()
-      this.locks.set(id, lock)
-      return lock
-    }
-
-    return await new Promise((resolve, reject) => {
-      // Using setImmediate to:
-      // 1. Prevent stack overflow by deferring recursive calls to the next event loop iteration.
-      // 2. Allow event loop to process other pending events, maintaining server responsiveness.
-      // 3. Ensure fairness in lock acquisition by giving other requests a chance to acquire the lock.
-      setImmediate(() => {
-        this.acquireLock(id, signal).then(resolve).catch(reject)
-      })
-    })
-  }
-
-  async unlock(id: string): Promise<void> {
-    const lock = this.locks.get(id)
-    if (!lock) {
-      throw new Error('Releasing an unlocked lock!')
-    }
-
-    this.locks.delete(id)
-  }
-
-  protected waitTimeout(signal: AbortSignal) {
-    return new Promise<void>((resolve) => {
-      const timeout = setTimeout(() => {
-        resolve()
-      }, this.timeout)
-
-      const abortListener = () => {
-        clearTimeout(timeout)
-        signal.removeEventListener('abort', abortListener)
-        resolve()
-      }
-      signal.addEventListener('abort', abortListener)
-    })
-  }
 }
