+++ packages/server/src/models/index.ts
@@ -3,4 +3,4 @@ export * as Metadata from './Metadata'
 export {StreamSplitter} from './StreamSplitter'
 export {Uid} from './Uid'
 export {Upload} from './Upload'
-export {Locker, MemoryLocker} from './Locker'
+export {Locker} from './Locker'
